services/requirements.runtime.txt
# Runtime-only dependencies for running a single bot container:
#   python -m services.bot_runtime
#
# NOTE: Keep this file free of API/orchestrator dependencies such as:
#   - fastapi, uvicorn
#   - docker (python sdk)
#
# Also intentionally excludes legacy "bot/" stack and heavy scientific deps (e.g. scipy).

# Settings / config
pydantic>=2.7
pydantic-settings>=2.3

# HTTP / WS / exchange
httpx>=0.27
websockets>=12.0
ccxt>=4.4

# DB access (runtime reads/writes engine snapshots)
SQLAlchemy>=2.0
asyncpg>=0.29
psycopg[binary]>=3.2

# Serialization
orjson>=3.10

# Optional: local/dev runs only (.env autoload in __main__.py). Orchestrator injects env in containers.
python-dotenv>=1.0

# Optional structured logging
structlog>=24.1
